adding shaper_calibrate_classic
